Не рабочий коммит

+++ src/main/java/bstu/BI/entity/domain/RentalTicket.java
@@ -12,12 +12,12 @@ import java.util.Date;
 public class RentalTicket {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
     @JsonIgnore
     @Column(nullable = false)
-    private Integer userId;
+    private Long userId;
     @Column(nullable = false)
-    private Integer bookTypeId;
+    private Long bookTypeId;
     @Column(nullable = false)
     private Double rentPrice;
     @Column(nullable = false)
