+++ src/main/java/bstu/BI/web/dto/DTO_Rent_RefundOfDeposit.java
@@ -9,5 +9,5 @@ public class DTO_Rent_RefundOfDeposit {
     @NotNull
     UserRequisites requisites;
     @NotNull
-    Integer bookTypeId;
+    Long bookTypeId;
 }
